Project/gen-ai-server/pyproject.toml
[project]
name = "gen-ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate",
    "av<14",
    "einops",
    "fastapi",
    "gpustat",
    "numpy>=1.26.0",
    "ollama",
    "pillow",
    "pydub",
    "pydantic",
    "pydantic-settings",
    "pytest",
    "python-dotenv",
    "python-multipart",
    "requests",
    "safetensors",
    "soundfile",
    "timm",
    "coqui-tts",
    "transformers",
    "uvicorn",
]

[tool.uv]
constraint-dependencies = [
    "numpy>=1.26.0",
    "ml-dtypes>=0.3.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
]

